consolidate escapeHtml into shared utility
Move duplicate escapeHtml function from commentsPanel.ts and
page-select-client.ts to src/utils/sanitize.ts
